requirements.txt
Django==1.5
djangorestframework==2.3.10
Sphinx==1.2
